Moves main functional control into a separate module so it can be required and use to build a global cli for bootstrapping a project template.
