+++ .gitignore
@@ -1,6 +1,7 @@
 node_modules
 .DS_Store
 coverage
+.idea
 
 # When the tests create files they are put here.
 tests/test-templates-output
