+++ playground/run.sh
@@ -1,3 +1,3 @@
-#! bin/bash
+#! /bin/bash
 node ../src/index.js
 # node ../src/index.js -c=custom-config.js
